cleaning up tab component styles
