--- src/apps/pages/Tabs.astro
@@ -1,83 +1,0 @@
----
-import _ from 'underscore';
-import TabsComponent from './Tabs';
-import PageBody from '@components/PageBody.astro';
-import clsx from 'clsx';
-import { toBackgroundClass } from '@utils/pageBuilder';
-
-const {
-  tabs,
-  raise,
-  activeBg,
-  textStyle,
-  invertText,
-  inactiveBg,
-  isTinaPreview
-} = Astro.props;
----
-
-{ !isTinaPreview && ( 
-  <TabsComponent tabs={tabs} raise={raise} activeBg={activeBg} textStyle={textStyle} invertText={invertText} inactiveBg={inactiveBg} client:only='react'>
-    {
-      _.map(tabs, (tab: any) => (
-        <div class='tab-content flex flex-col' id={`${tab.label}--content`}>
-          <PageBody sections={tab.content} />
-        </div>
-      ))
-    }
-  </TabsComponent>
-)}
-
-{
-  isTinaPreview && (
-    <div>
-      <div class={clsx(
-        'px-6 sm:px-12 md:px-16 lg:px-32 2xl:mx-auto max-w-(--breakpoint-2xl) overflow-x-auto overflow-y-visible',
-      )}>
-        <div aria-label='Tabs' class={clsx(
-          'flex min-h-[65px] relative z-30',       
-          { 'w-full border-b-4 h-[65px]': !raise }
-        )}>
-          {
-            _.map(tabs, (tab: any, idx: number) => ( 
-              <button 
-                class={clsx(
-                  'tab-button',
-                  'lg:whitespace-nowrap px-3 lg:px-6 xl:px-8 py-4 focus:outline-none',
-                  { '-mb-[4px] border-b-4': !raise },
-                  idx === 0 && toBackgroundClass(activeBg),
-                  idx !== 0 && toBackgroundClass(inactiveBg),
-                  (idx !== 0 || !toBackgroundClass(activeBg)) && raise && invertText && 'text-text-inverse',
-                  { 'hover:border-secondary/60 border-b-4 hover:text-secondary/60': !raise },
-                  { 'border-secondary border-b-4 text-secondary': idx === 0 && !raise },
-                  { 'cursor-default': idx === 0 },
-                  { 'font-serif italic sm:text-lg lg:text-xl font-normal': textStyle === 'italic' },
-                  { 'text-sm text-gray-500 uppercase font-medium': textStyle === 'uppercase' }
-                )}
-                id={`tab-${idx}--button`}
-              >
-                { tab.label }
-              </button>
-            ))
-          }
-        </div>
-      </div>
-      <div>
-      {
-        _.map(tabs, (tab: any, idx: number) => (
-          <div 
-            class={clsx(
-              'tab-content flex flex-col',
-              )} 
-            id={`tab-${idx}--content`}
-          >
-            <h3 class="w-full text-center italic border-b border-b-black text-lg">{tab.label} Content Preview</h3>
-            <PageBody sections={tab.content} isTinaPreview />
-          </div>
-        ))
-      }
-      </div>
-    </div>
-  )
-}
-
