+++ src/components/PageBody.astro
@@ -8,7 +8,7 @@ import Images from '@apps/pages/Images.astro';
 import LinksBanner from '@apps/pages/LinksBanner.astro';
 import MultiColumn from '@apps/pages/MultiColumn.astro';
 import Spacer from '@components/Spacer.astro';
-import Tabs from '@apps/pages/Tabs.astro';
+import Tabs from '@apps/pages/Tabs/index.astro';
 import TextImageReact from '@apps/pages/TextImage';
 import TextImageAstro from '@apps/pages/TextImage.astro';
 import Container from '@components/Container.astro';
